Change to int and update preconditions check

+++ dialogue-clients/src/main/java/com/palantir/dialogue/clients/ChannelCache.java
@@ -288,7 +288,7 @@ final class ChannelCache {
 
         boolean concurrencyLimiterExponentialRamp();
 
-        double concurrencyLimiterExponentialRampInitialLimit();
+        int concurrencyLimiterExponentialRampInitialLimit();
     }
 
     @Unsafe
