+++ dialogue-clients/src/main/java/com/palantir/dialogue/clients/DialogueClients.java
@@ -243,7 +243,7 @@ public final class DialogueClients {
         ReloadingFactory withConcurrencyLimiterExponentialRamp(boolean enabled);
 
         @Beta
-        ReloadingFactory withConcurrencyLimiterExponentialRamp(boolean enabled, double initialLimit);
+        ReloadingFactory withConcurrencyLimiterExponentialRamp(boolean enabled, int initialLimit);
 
         StickyChannelFactory getStickyChannels(String serviceName);
